feat(internal/cmd): add an alternate cli
This adds an alternate CLI that, in my opinion, has a cleaner interface.
If no argument is given, the REPL is used. If an argument is given, the
file is run through the normal query path which would give the
equivalent answer as InfluxDB's query endpoint.

This CLI also supports sending a jaeger trace.