chore: remove libflux site
This directory seems to be an integration to work out wasm work more
than two years ago. There's no scaffolding currently to generate a wasm
binary that can be used by this site, and it's a common place for
dependabot to keep putting patches up to update dependencies. Rather
than trying to keep dependencies up to date, I think we should just
delete the whole thing.